feat(common): resolve pr problem
